Implemented single-source version
Following the advice found here: https://packaging.python.org/en/latest/guides/single-sourcing-package-version/

I have implemented option #3 to achieve a version number that only has
to be set in one place. This is in a new file called _version.py.

- The package imports it
- The setup.py executes the file and parses the version number that way

To do it otherwise results in setup.py importing the package before it's
fully installed and can be extremely problematic so a method like this
is recommended.