script to auto-made generate template structures
Added script to root of repo (it shouldn't be included in any build and
should not otherwise be available to users) that when triggered builds a
new markdown file in the source of the docs, with details of all the
current templates available.

Added a tigger for this when the docs are built.