project name is too long so switch to abbrev
the full project name is too long and needs to be shortened. In the
absence of good contractions I've gone for a simple acronym of acpp.

Additionally removes complication of confusing "package" with "project"
which I do every time I type it out.

This change does not change the name in pypi, it is just the package
name when running it i.e. python -m acpp.