Added entry points to setup.py
Pipx was failing because we added the alias of acpp for pure pip users,
which also broke pipx (which needs the package and the run command to be
the same).

This fix implements entry points to setup.py. Now pip users do not even
need to specify the "python -m" and can just invoke "acpp --help"
directly. But it re-adds the capability to run the tool with
"ansys-create-python-project", as pipx requires. I have not been able to
test this beyond a local pipx execution (which worked). We need to push
to PyPI first. It might be worth pushing to testpypi? I'm not sure. It
can be tricky getting pipx to work with that. Given our userbase is
relativeky small, it might just be OK to push to PyPI and fix it up
relatively quickly.

I have also altered the documentation to represent what should be
working and the new syntax.